separate SSR utils from utils.js
